Added UsbEcho command. Changed KosDropSeconds from int32 to double to allow drop times less than a second. hh 9/28/25

+++ src/AutostartKos.c
@@ -58,10 +58,10 @@ void AutostartKos(double MsLevel){
     if(0==KosLockout){                 // KOS not locked out by receive data
       if(OldLoopSense!=LOOP_SENSE_Get()){ // Loop changed, so key up or stay keyed up - Don't need to check NoLoop since looking for change instead of space. 
         OldLoopSense=LOOP_SENSE_Get();    // Remember new value
-        KosCounter=8000*UserConfig.KosDropSeconds;  // Reset counter to drop out later
+        KosCounter=(uint32_t)(8000.0*UserConfig.KosDropSeconds);  // Reset counter to drop out later
       }
       if(0==BaudotUartTxOut){       // Set KOS counter on BaudotUart space
-        KosCounter=8000*UserConfig.KosDropSeconds;  // Reset counter to drop out later 
+        KosCounter=(uint32_t)(8000.0*UserConfig.KosDropSeconds);  // Reset counter to drop out later 
       }
       if(0!=KosCounter) KosCounter--;   // Decrement towards zero
       if(KosCounter!=0){              // Transmit if not timed out
